refactor(auth): restructure authentication logic and route management
- Removed obsolete route guards and matchers from the auth module.
- Consolidated route configuration into a single file for better maintainability.
- Updated import paths for the AuthProvider and related utilities.
- Enhanced route validation logic to streamline access control based on user roles.
- Implemented a new AuthProvider component to manage authentication state and user sessions.
- Added utility functions for handling JWT tokens and user authentication checks.
- Refactored components to utilize the new AuthProvider and updated context.
