+++ frontend/src/community/auth/enums/auth.ts
@@ -1,0 +1,10 @@
+export enum AuthMethods {
+  CREDENTIAL = "CREDENTIAL",
+  GOOGLE = "GOOGLE",
+  MICROSOFT = "MICROSOFT"
+}
+
+export enum SignInStatus {
+  SUCCESS = "SUCCESS",
+  FAILURE = "FAILURE"
+}
