+++ frontend/src/community/common/components/templates/ContentLayout/ContentLayout.tsx
@@ -11,9 +11,9 @@ import Head from "next/head";
 import { useRouter } from "next/router";
 import { JSX, memo, useEffect, useMemo } from "react";
 
+import { useAuth } from "~community/auth/providers/AuthProvider";
 import { useGetOrganization } from "~community/common/api/OrganizationCreateApi";
 import { useStorageAvailability } from "~community/common/api/StorageAvailabilityApi";
-import { useAuth } from "~community/common/auth/AuthProvider";
 import Button from "~community/common/components/atoms/Button/Button";
 import Icon from "~community/common/components/atoms/Icon/Icon";
 import VersionUpgradeBanner from "~community/common/components/molecules/VersionUpgradeBanner/VersionUpgradeBanner";
