+++ frontend/src/community/common/hooks/useOrgSetupRedirect.tsx
@@ -1,17 +1,17 @@
-import { useSession } from "next-auth/react";
 import { useRouter } from "next/router";
 
 import ROUTES from "../constants/routes";
 import { OrgSetupStatusType } from "../types/OrganizationCreateTypes";
+import { useAuth } from "~community/auth/providers/AuthProvider";
 
 const useOrgSetupRedirect = () => {
   const router = useRouter();
 
-  const { status } = useSession();
+  const { isAuthenticated } = useAuth();
 
   let isSignInSessionAvailable = false;
 
-  if (status === "authenticated") {
+  if (isAuthenticated) {
     isSignInSessionAvailable = true;
   }
   const navigateByStatus = (orgSetupStatus: OrgSetupStatusType) => {
